ExamAutoPro/requirements.txt
# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.1.0

# AI/ML Libraries
opencv-python==4.8.1.78
face-recognition==1.3.0
pytesseract==0.3.10
scikit-learn==1.3.2
nltk==3.8.1

# Natural Language Processing
spacy==3.7.2
textblob==0.17.1

# Audio Processing
librosa==0.10.1

# Web Scraping (for content analysis)
beautifulsoup4==4.12.2
requests==2.31.0

# File Handling
python-magic==0.4.27
pdfplumber==0.11.9
pymupdf==1.27.2.2
ocrmypdf==17.4.1
pdf2image==1.17.0

# Security
cryptography>=41.0.0

# Development Tools
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0

# Email
django-anymail==10.2

# Caching
redis==5.0.1
django-redis==5.4.0

# Task Queue
celery==5.3.4

# Monitoring
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Production
uvicorn==0.24.0

# Optional handwriting OCR (Google Lens-like)
google-cloud-vision
